backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
pymongo==4.5.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
bcrypt==4.0.1
googletrans==4.0.0rc1
twilio==8.8.0
python-dateutil==2.8.2
gunicorn==21.2.0 